logger/requirements.txt
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
